Restructure project and add README
