+++ pkg/jsonresolver/resolver.go
@@ -7,7 +7,7 @@ import (
 	"strings"
 
 	"github.com/Knetic/govaluate"
-	"github.com/example/user-config-resolver-go/api"
+	"github.com/example/user-config-resolver-go/pkg/resolver"
 )
 
 type JsonConfigResolverService struct {
@@ -20,14 +20,14 @@ func (s *JsonConfigResolverService) SetConfigToResolve(config string) { s.config
 
 func (s *JsonConfigResolverService) ResolveConfig(groups []string) (string, error) {
 	if s.configToResolve == "" {
-		return "", api.ConfigResolverError{fmt.Errorf("config to resolve is empty")}
+		return "", resolver.ConfigResolverError{fmt.Errorf("config to resolve is empty")}
 	}
 	return s.ResolveConfigFrom(s.configToResolve, groups)
 }
 
 func (s *JsonConfigResolverService) ResolveConfigInto(groups []string, target any) error {
 	if s.configToResolve == "" {
-		return api.ConfigResolverError{fmt.Errorf("config to resolve is empty")}
+		return resolver.ConfigResolverError{fmt.Errorf("config to resolve is empty")}
 	}
 	return s.ResolveConfigFromInto(s.configToResolve, groups, target)
 }
@@ -45,12 +45,12 @@ func (s *JsonConfigResolverService) ResolveConfigFromInto(cfg string, groups []s
 	dec := json.NewDecoder(strings.NewReader(cfg))
 	dec.DisallowUnknownFields()
 	if err := dec.Decode(&c); err != nil {
-		return api.ConfigResolverError{err}
+		return resolver.ConfigResolverError{err}
 	}
 	applyRules(groups, &c)
 	data, err := json.Marshal(c.DefaultProperties)
 	if err != nil {
-		return api.ConfigResolverError{err}
+		return resolver.ConfigResolverError{err}
 	}
 	if strPtr, ok := target.(*string); ok {
 		*strPtr = string(data)
