+++ pkg/jsonresolver/testdata/custom-user-groups/input.json
