+++ pkg/jsonresolver/testdata/custom-user-groups/output.json
