+++ pkg/jsonresolver/testdata/invalid-config/input.json
