+++ pkg/jsonresolver/testdata/user-in-all-groups/input.json
