+++ pkg/jsonresolver/testdata/user-in-all-groups/output.json
