+++ pkg/jsonresolver/testdata/user-in-any-groups/input.json
