+++ pkg/jsonresolver/testdata/user-in-any-groups/output.json
