+++ pkg/jsonresolver/testdata/user-in-different-groups/input.json
