+++ pkg/jsonresolver/testdata/user-in-different-groups/output.json
