+++ pkg/jsonresolver/testdata/user-in-no-groups/input.json
