+++ pkg/jsonresolver/testdata/user-in-no-groups/output.json
